+++ src/http/controllers/images-controller.ts
@@ -1,7 +1,6 @@
 import type { FastifyReply, FastifyRequest } from 'fastify'
 import { createImageQuerySchema } from '../schemas/images'
-import { deleteOldImagesService } from '@/domain/services/r2-storage/delete-old-images'
-import { uploadImageService } from '@/domain/services/r2-storage/upload-image'
+import { createCloudStorage } from '@/factories/cloud-storage-factory'
 
 const MAXIMUM_FILE_SIZE_IN_BYTES = 1024 * 1024 * 4 // 4mb
 
@@ -18,12 +17,14 @@ export async function createImageController(
     return reply.status(400).send({ message: 'Invalid file provided.' })
   }
 
+  const cloudStorage = createCloudStorage('R2')
+
   const { file: contentStream, mimetype } = uploadedFile
 
   const timestamp = Date.now()
-  await deleteOldImagesService(`${folder}/${request.user.id}`)
+  await cloudStorage.deleteOldImages(`${folder}/${request.user.id}`)
 
-  const { url } = await uploadImageService({
+  const { url } = await cloudStorage.uploadImage({
     path: `${folder}/${fileName || request.user.id}-${timestamp}`,
     contentStream,
     contentType: mimetype,